requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn==1.0.2
umap-learn==0.5.3
plotly>=5.3.0
xgboost>=1.4.0
networkx>=2.6.0 